+++ src/app/helpers/index.ts
@@ -3,3 +3,4 @@ export * from './dialog.helper';
 export * from './filesystem.helper';
 export * from './inflector.helper';
 export * from './placeholder.helper';
+export * from './validation-helpers';